PR review addressed

+++ feature/gribi/otg_tests/backup_nhg_multiple_nh_test/backup_nhg_multiple_nh_test.go
@@ -316,13 +316,13 @@ func validateTrafficFlows(t *testing.T, ate *ondatra.ATEDevice, ateTop gosnappi.
 	otgutils.LogFlowMetrics(t, ate.OTG(), ateTop)
 	otgutils.LogPortMetrics(t, ate.OTG(), ateTop)
 	recvMetric := gnmi.Get(t, ate.OTG(), gnmi.OTG().Flow(flow).State())
-	txPackets := recvMetric.GetCounters().GetOutPkts()
-	rxPackets := recvMetric.GetCounters().GetInPkts()
+	txPackets := float32(recvMetric.GetCounters().GetOutPkts())
+	rxPackets := float32(recvMetric.GetCounters().GetInPkts())
 	lostPackets := txPackets - rxPackets
 	if txPackets == 0 {
 		t.Fatalf("Tx packets should be higher than 0")
 	}
-	got := float32(lostPackets * 100 / txPackets)
+	got := lostPackets * 100 / txPackets
 	if drop {
 		if got != 100 {
 			t.Fatalf("Traffic passing for flow %s got %f, want 100 percent loss", flow, got)
